+++ src/kcp/error.rs
@@ -6,8 +6,6 @@ use std::io::{self, ErrorKind};
 pub enum Error {
     #[error("conv inconsistent, expected {0}, found {1}")]
     ConvInconsistent(u32, u32),
-    #[error("invalid mtu {0}")]
-    InvalidMtu(usize),
     #[error("invalid segment size {0}")]
     InvalidSegmentSize(usize),
     #[error("invalid segment data size, expected {0}, found {1}")]
@@ -45,7 +43,6 @@ impl From<Error> for io::Error {
             Error::Io(err) => return err,
             Error::RecvQueueEmpty | Error::ExpectingFragment => ErrorKind::WouldBlock,
             Error::ConvInconsistent(..)
-            | Error::InvalidMtu(..)
             | Error::InvalidSegmentSize(..)
             | Error::SegmentDataSizeMismatch(..)
             | Error::NeedUpdate